+++ indico/modules/events/editing/notifications.py
@@ -22,7 +22,6 @@ def notify_comment(comment):
     """Notify about a new comments on a revision."""
     revision = comment.revision
     editable = revision.editable
-    contribution = editable.contribution
     editor = editable.editor
     submitter = next((r.user for r in editable.revisions[::-1] if r.is_submitter_revision), None)
     author = comment.user
@@ -48,8 +47,7 @@ def notify_comment(comment):
                                       author_name=author_name,
                                       timeline_url=editable.external_timeline_url,
                                       recipient_name=recipient.first_name,
-                                      contribution_title=contribution.title,
-                                      program_code=contribution.code,
+                                      contribution=editable.contribution,
                                       text=comment.text)
             email = make_email(recipient.email, template=tpl)
         send_email(email, editable.event, 'Editing', log_metadata={'editable_id': editable.id})