+++ CHANGES.rst
@@ -63,6 +63,8 @@ Improvements
 - Do not bother people registering using an invitation link with a CAPTCHA (:pr:`6095`)
 - Add option to allow people to register using an invitation link even if the event is
   restricted (:pr:`6094`)
+- Include comment and contribution title on timeline notifications to editors/authors
+  (:issue:`6027`, :pr:`6042`)
 
 Bugfixes
 ^^^^^^^^